refactor(runtime): extract regex + declaration types out of mod.rs
`runtime/mod.rs` had grown to 2470 lines, far over the 500-line file
guideline, dominated by type definitions that have nothing to do with
the (irreducible ~950-line) `Interpreter` struct that has to live there.

Move two cohesive, self-contained type clusters into their own modules,
re-exported from `runtime` so every existing `crate::runtime::T` /
`use super::*` reference keeps resolving unchanged:

- `regex_types.rs` — the parsed-pattern model (`RegexPattern`/`RegexToken`/
  `RegexAtom`/`RegexQuant`/`CharClass`/`ClassItem`/`RegexSeparatorSpec`),
  the capture accumulator (`RegexCaptures`/`CodeBlockContext`/
  `QuantifiedCaptureEntry`/`SILENT_ACTION_MARKER_PREFIX`). These belong
  next to the regex engine, not in the interpreter's mod.rs.
- `decl_types.rs` — the class/role/method declaration metadata
  (`ClassDef`/`RoleDef`/`RoleCandidateDef`/`SubsetDef`/`MethodDef`/
  `ProtoMethodCtx`/`MethodDispatchFrame`/`WrapDispatchFrame`/
  `SquishIteratorMeta`/`MultiDispatchEntry`).

Previously module-private structs and their fields (read directly by the
sibling `regex`/`class`/`dispatch` modules) are widened to `pub(crate)`;
this is a visibility widening only, no behavior change. `mod.rs` drops
from 2470 to 2080 lines.

Pure code movement — build + `cargo clippy -- -D warnings` + `make test`
(Files=2083, Tests=19703) all green.

Also record 2026-07-21 perf-scouting findings in PLAN.md §5: item 1
(SipHash removal) is already DONE (`CompiledFns` is
`FxHashMap<Symbol, ...>`); item 2 (callsite-line marker) and item 0
(needs_env_sync blanket) are documented as deferred with the reasons.
