+++ src/runtime/mod.rs
@@ -172,6 +172,8 @@ mod class;
 mod class_dispatch;
 mod class_introspection;
 pub(crate) use class_introspection::UserMethodOrAccessor;
+mod decl_types;
+pub(crate) use self::decl_types::*;
 pub(crate) mod deprecation;
 pub(crate) mod did_you_mean;
 mod dispatch;
@@ -294,6 +296,7 @@ mod regex_parse_charclass;
 mod regex_parse_core;
 mod regex_parse_ltm;
 mod regex_parse_modifier;
+mod regex_types;
 mod registration;
 mod registration_class;
 mod registration_class_attr;
@@ -356,6 +359,7 @@ pub(crate) mod value_iterator;
 pub(crate) use self::output_sink::OutputSink;
 #[allow(unused_imports)]
 pub(crate) use self::output_sink::{OutputSinkReadGuard, OutputSinkWriteGuard};
+pub(crate) use self::regex_types::*;
 pub(crate) use self::registration_class::ClassDeclModifiers;
 pub(crate) use self::registry::Registry;
 pub(crate) use self::tap_state::{TapState, TestState, TodoRange};
@@ -501,156 +505,6 @@ pub(crate) fn sym_mro(names: &[&str]) -> std::sync::Arc<[crate::symbol::Symbol]>
         .iter()
         .map(|s| crate::symbol::Symbol::intern(s))
         .collect()
-}
-
-#[derive(Clone, Default)]
-pub(crate) struct ClassDef {
-    parents: Vec<String>,
-    // (name, is_public, default, is_rw, is_required, sigil, where_constraint)
-    attributes: Vec<ClassAttributeDef>,
-    attribute_types: HashMap<String, String>, // attr_name -> type constraint
-    attribute_smileys: HashMap<String, String>, // attr_name -> smiley ("D", "U", "_")
-    attribute_built: HashMap<String, bool>,
-    /// Attributes declared with `has $x` (no twigil) — the bare name is an alias
-    /// for `$!x` inside class methods.
-    alias_attributes: HashSet<String>,
-    methods: HashMap<String, Vec<MethodDef>>, // name -> overloads
-    native_methods: HashSet<String>,
-    mro: std::sync::Arc<[crate::symbol::Symbol]>,
-    /// Attribute var names (e.g. "!foo") that have `handles *` wildcard delegation.
-    wildcard_handles: Vec<String>,
-    /// Class-level attributes declared with `our $.x` or `my $.x` (shared across instances).
-    /// Maps attribute name to its current value.
-    class_level_attrs: HashMap<String, Value>,
-}
-
-#[derive(Debug, Clone)]
-pub(crate) struct RoleDef {
-    pub(crate) attributes: Vec<ClassAttributeDef>,
-    pub(crate) methods: HashMap<String, Vec<MethodDef>>,
-    pub(crate) is_stub_role: bool,
-    pub(crate) is_hidden: bool,
-    /// Whether this role was declared with `is rw` or `also is rw`.
-    /// Used during `register_role_decl` to compute effective is_rw for attributes.
-    #[allow(dead_code)]
-    pub(crate) is_rw: bool,
-    /// Captured environment for evaluating attribute defaults in closures.
-    pub(crate) captured_env: Option<HashMap<String, Value>>,
-    /// Attribute var names (e.g. "!foo") that have `handles *` wildcard delegation.
-    pub(crate) wildcard_handles: Vec<String>,
-    /// Unique identifier for this role definition instance, used to distinguish
-    /// different lexical roles with the same name.
-    pub(crate) role_id: u64,
-    /// Attribute conflicts detected during role-to-role composition.
-    /// Each entry is (attr_name, declaring_role, conflicting_role).
-    pub(crate) attribute_conflicts: Vec<(String, String, String)>,
-    /// Attribute names declared directly in this role's body (not inherited
-    /// via `does`). Used to disambiguate diamond composition (where the same
-    /// attribute reaches via two paths from a shared ancestor) from a real
-    /// attribute conflict.
-    pub(crate) own_attribute_names: HashSet<String>,
-    /// Body statements deferred until composition time (for parameterized roles).
-    /// These are non-method/non-attribute statements that may reference type parameters
-    /// and must be re-executed for each class composition with concrete type bindings.
-    pub(crate) deferred_body_stmts: Vec<Stmt>,
-    /// Unknown lowercase trait names deferred for custom `trait_mod:<is>` dispatch.
-    pub(crate) deferred_custom_traits: Vec<String>,
-}
-
-#[derive(Debug, Clone)]
-pub(crate) struct RoleCandidateDef {
-    type_params: Vec<String>,
-    type_param_defs: Vec<ParamDef>,
-    role_def: RoleDef,
-    /// Parent classes/roles declared via `is` on this candidate.
-    parents: Vec<String>,
-    /// Language version (e.g. "6.c") captured at registration time.
-    language_version: String,
-}
-
-#[derive(Debug, Clone)]
-pub(crate) struct SubsetDef {
-    base: String,
-    predicate: Option<Expr>,
-    version: String,
-}
-
-#[derive(Debug, Clone)]
-pub(crate) struct MethodDef {
-    pub(crate) params: Vec<String>,
-    pub(crate) param_defs: Vec<ParamDef>,
-    /// Method body AST. Wrapped in Arc to make MethodDef clones O(1) since
-    /// the body is never mutated after construction and can be large.
-    pub(crate) body: std::sync::Arc<Vec<Stmt>>,
-    pub(crate) is_rw: bool,
-    pub(crate) is_private: bool,
-    pub(crate) is_multi: bool,
-    pub(crate) is_my: bool,
-    /// Role where this method was originally declared when composed into a class.
-    pub(crate) role_origin: Option<String>,
-    /// The deepest/original role where this method was first defined (for diamond detection).
-    pub(crate) original_role: Option<String>,
-    pub(crate) return_type: Option<String>,
-    pub(crate) compiled_code: Option<std::sync::Arc<crate::opcode::CompiledCode>>,
-    /// Delegation info: (attribute_var_name, target_method_name).
-    /// When set, the method forwards the call (with all args) to the named method
-    /// on the object stored in the given attribute.
-    pub(crate) delegation: Option<(String, String)>,
-    /// `is default` trait — this candidate is preferred when multi dispatch ties.
-    pub(crate) is_default: bool,
-    /// `is DEPRECATED` message: None = not deprecated.
-    pub(crate) deprecated_message: Option<String>,
-    /// Whether this is a submethod (not inherited by subclasses).
-    pub(crate) is_submethod: bool,
-}
-
-/// Invocant context for an active `proto method` `{*}` dispatch.
-#[derive(Debug, Clone)]
-pub(crate) struct ProtoMethodCtx {
-    pub(crate) invocant: Value,
-}
-
-/// One entry of `multi_dispatch_stack`: (function_name, remaining_candidates,
-/// original_args, first_candidate_rw_params). See the field doc on
-/// `Interpreter::multi_dispatch_stack`.
-type MultiDispatchEntry = (
-    String,
-    Vec<Arc<FunctionDef>>,
-    Vec<Value>,
-    Vec<(usize, String)>,
-);
-
-#[derive(Debug, Clone)]
-struct MethodDispatchFrame {
-    receiver_class: String,
-    invocant: Value,
-    args: Vec<Value>,
-    remaining: Vec<(String, MethodDef)>,
-    /// The FIRST (winning) candidate's scalar `is rw`/`is raw` positional params
-    /// as (positional_arg_index, sigil-less_param_name). Stays fixed across the
-    /// MRO chain so a `nextsame`+rw redispatch can forward the rw param's current
-    /// value and route the next candidate's writeback through it (§D capstone).
-    rw_params: Vec<(usize, String)>,
-}
-
-/// Frame for navigating through wrapper chain during callsame/callwith.
-#[derive(Debug, Clone)]
-pub(crate) struct WrapDispatchFrame {
-    /// The sub id being wrapped (to prevent re-entrant wrap dispatch).
-    pub(crate) sub_id: u64,
-    /// Remaining callables: inner wrappers then original sub. Next to call is first.
-    pub(crate) remaining: Vec<Value>,
-    /// Original call arguments.
-    pub(crate) args: Vec<Value>,
-}
-
-#[derive(Debug, Clone)]
-pub(crate) struct SquishIteratorMeta {
-    pub(crate) source_items: Vec<Value>,
-    pub(crate) as_func: Option<Value>,
-    pub(crate) with_func: Option<Value>,
-    pub(crate) revert_values: HashMap<String, Value>,
-    pub(crate) revert_remove: Vec<String>,
 }
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
@@ -816,256 +670,6 @@ pub(crate) struct IoHandleState {
     /// When set, the handle is closed automatically the moment word iteration
     /// reaches EOF (Raku's `words($fh, :close)` close-on-exhaust semantics).
     close_on_word_exhaust: bool,
-}
-
-#[derive(Clone)]
-struct RegexPattern {
-    tokens: Vec<RegexToken>,
-    anchor_start: bool,
-    anchor_end: bool,
-    ignore_case: bool,
-    ignore_mark: bool,
-}
-
-/// Context stored for each code block encountered during regex matching.
-#[derive(Clone)]
-pub(crate) struct CodeBlockContext {
-    pub(crate) code: String,
-    pub(crate) named: HashMap<String, Vec<String>>,
-    pub(crate) matched_so_far: String,
-    pub(crate) positional: Vec<String>,
-}
-
-/// A single entry in a quantified capture list: (matched_text, from, to, subcaptures).
-///
-/// The nested sub-captures are held behind an `Arc` so that cloning a parent
-/// `RegexCaptures` during backtracking is a refcount bump rather than a deep
-/// copy of the whole sub-match tree. A completed sub-match is effectively
-/// immutable once stored; the rare post-store tweak (e.g. setting `action_name`)
-/// goes through `Arc::make_mut`, which is free while the entry is still unshared.
-pub(crate) type QuantifiedCaptureEntry = (String, usize, usize, Option<Arc<RegexCaptures>>);
-
-/// Prefix marking a `named_subcaps` entry as a *silent action capture*: the
-/// match of a silent subrule (`<.foo>`) that is hidden from `.hash` but whose
-/// grammar action method (and its descendants') must still fire. The prefix is a
-/// control character that can never appear in a real capture name, so marker
-/// entries never collide with user captures and are trivially filtered.
-pub(crate) const SILENT_ACTION_MARKER_PREFIX: &str = "\u{1}silent\u{1}";
-
-#[derive(Clone, Default)]
-pub(crate) struct RegexCaptures {
-    pub(crate) named: HashMap<String, Vec<String>>,
-    /// Nested sub-captures for named subrule matches. Key is capture name,
-    /// value is inner captures from the subrule (parallel to entries in `named`).
-    pub(crate) named_subcaps: HashMap<String, Vec<Arc<RegexCaptures>>>,
-    pub(crate) positional: Vec<String>,
-    /// Nested sub-captures for positional capture groups. Each entry corresponds
-    /// to the same index in `positional` and holds inner captures from nested groups.
-    pub(crate) positional_subcaps: Vec<Option<Arc<RegexCaptures>>>,
-    /// When a capture group is quantified (e.g. `(\w)+`), this parallel vec
-    /// stores the list of all iteration matches for that positional index.
-    /// When `Some`, the positional slot should be rendered as an Array of Match objects.
-    pub(crate) positional_quantified: Vec<Option<Vec<QuantifiedCaptureEntry>>>,
-    /// Character offsets (start, end) for each entry in `positional`.
-    pub(crate) positional_offsets: Vec<(usize, usize)>,
-    /// Marks a positional slot as an *unmatched optional* capture (`(x)?` that
-    /// matched zero times) which must render as `Nil` (not an empty Match). Only
-    /// the zero-match reservation arms set entries here; they pad with `false` up
-    /// to the current `positional` length before pushing `true`, so matched
-    /// captures (which never touch this vec) stay aligned. The Match builder reads
-    /// it via `.get(i)` — a missing/`false` entry renders normally.
-    pub(crate) positional_nil: Vec<bool>,
-    /// Unnamed capture slots by capture index (for $0, $1, ...), where `None`
-    /// represents an unmatched capture.
-    pub(crate) positional_slots: Vec<Option<(String, usize, usize)>>,
-    pub(crate) matched: String,
-    pub(crate) from: usize,
-    pub(crate) to: usize,
-    pub(crate) capture_start: Option<usize>,
-    pub(crate) capture_end: Option<usize>,
-    /// Starting position of the match in the input (character index).
-    /// Set at the beginning of regex matching to allow code blocks to compute
-    /// the matched-so-far text.
-    pub(crate) match_from: usize,
-    /// Code blocks encountered during matching (code + captures at that point).
-    /// Executed after match for side effects.
-    /// Each entry: (code, named_captures, matched_so_far, positional_captures)
-    pub(crate) code_blocks: Vec<CodeBlockContext>,
-    /// Variables declared via `:my $var = expr;` inside regex.
-    /// These are made available to `<{ code }>` closures.
-    pub(crate) regex_vars: HashMap<String, Value>,
-    /// The winning :sym<> variant name, if this match was from a protoregex.
-    pub(crate) sym: Option<String>,
-    /// Named captures from quantified tokens — always stored as arrays in Match.
-    pub(crate) named_quantified: HashSet<String>,
-    /// For aliased captures like `<str=.str_escape>`, maps capture name to
-    /// original rule name for grammar action dispatch.
-    pub(crate) capture_alias_map: HashMap<String, String>,
-    /// The original rule name when this capture was stored under an alias.
-    pub(crate) action_name: Option<String>,
-    /// Hash captures from `%<name>=(...)` aliasing in regex.
-    pub(crate) hash_captures: HashMap<String, Vec<(String, Option<String>)>>,
-}
-
-#[derive(Clone)]
-struct RegexToken {
-    atom: RegexAtom,
-    quant: RegexQuant,
-    named_capture: Option<String>,
-    /// Secondary named capture for capturing subrule aliases like `$<alias>=<builtin_class>`.
-    /// When set, the matched text is also stored under this name (the original rule name).
-    secondary_named_capture: Option<String>,
-    /// Hash aliasing: `%<name>=(...)` captures build a hash
-    hash_capture: Option<String>,
-    /// Array-sigil capture alias (`@<name>=(...)`): forces the named capture
-    /// into list context, so even a single (non-quantified) match yields a
-    /// one-element List rather than a bare Match. Mirrors Raku's `@`-sigil
-    /// declaration semantics for hypothetical capture variables.
-    force_list_capture: bool,
-    ratchet: bool,
-    /// Frugal (non-greedy) quantifier modifier: `*?`, `+?`, `??`
-    frugal: bool,
-    /// Separator for `%` / `%%` quantifiers, e.g. `<thing> +% ','`. When present,
-    /// the quantified atom is matched with `separator.atom` interleaved between
-    /// iterations. `allow_trailing` is true for `%%` (an optional trailing
-    /// separator is permitted). The separator's own captures are appended as
-    /// positional/named captures after the main atom's, matching Raku semantics.
-    separator: Option<Box<RegexSeparatorSpec>>,
-}
-
-#[derive(Clone)]
-struct RegexSeparatorSpec {
-    /// The separator sub-pattern (matched between iterations). Holding a full
-    /// pattern preserves named captures, quantifiers, and other structure of
-    /// complex separators such as `$<delim>=<[a..z]>*`.
-    pattern: RegexPattern,
-    allow_trailing: bool,
-}
-
-#[derive(Clone)]
-enum RegexAtom {
-    Literal(char),
-    Named(String),
-    Any,
-    CharClass(CharClass),
-    /// `<.ws>` — Raku's word-boundary-aware whitespace rule:
-    /// requires `\s+` between word characters, `\s*` otherwise.
-    WsRule,
-    Newline,
-    NotNewline,
-    Group(RegexPattern),
-    CaptureGroup(RegexPattern),
-    Alternation(Vec<RegexPattern>),
-    SequentialAlternation(Vec<RegexPattern>),
-    /// Conjunction: all branches must match at the same position; longest match wins
-    Conjunction(Vec<RegexPattern>),
-    ZeroWidth,
-    CodeAssertion {
-        code: String,
-        negated: bool,
-        is_assertion: bool,
-    },
-    /// `<{ code }>` — closure interpolation: evaluate code and match result as regex
-    ClosureInterpolation {
-        code: String,
-    },
-    UnicodeProp {
-        name: String,
-        negated: bool,
-        args: Option<String>,
-    },
-    UnicodePropAssert {
-        name: String,
-        negated: bool,
-    }, // zero-width assertion
-    CaptureStartMarker,
-    CaptureEndMarker,
-    /// `:my $var = expr;` — variable declaration inside a regex
-    VarDecl {
-        code: String,
-    },
-    /// Combined character class: <+ xdigit - lower>, matches positive AND NOT negative
-    CompositeClass {
-        positive: Vec<ClassItem>,
-        negative: Vec<ClassItem>,
-    },
-    /// Lookaround assertion: <?before pattern>, <!before pattern>,
-    /// <?after pattern>, <!after pattern>
-    Lookaround {
-        pattern: RegexPattern,
-        negated: bool,
-        is_behind: bool,
-    },
-    /// `<<` or `«` — left word boundary assertion (zero-width)
-    LeftWordBoundary,
-    /// `>>` or `»` — right word boundary assertion (zero-width)
-    RightWordBoundary,
-    /// `<?wb>` (word boundary) / `<!wb>` (not a word boundary) — zero-width
-    /// assertion at a transition between a word char and a non-word char (either
-    /// direction), i.e. `<<` or `>>`.
-    WordBoundary {
-        negated: bool,
-    },
-    /// `^^` — start of line assertion (zero-width)
-    StartOfLine,
-    /// `$$` — end of line assertion (zero-width)
-    EndOfLine,
-    /// `$0`, `$1`, etc. — backreference to positional capture group
-    Backref(usize),
-    /// `$<name>` — backreference to named capture group
-    NamedBackref(String),
-    /// `<?same>` / `<!same>` — zero-width assertion: adjacent chars are same/different
-    SameAssertion {
-        negated: bool,
-    },
-    /// `<at(N)>` — zero-width assertion: match at position N
-    AtPosition(usize),
-    /// Internal marker used while rewriting `left ~ goal inner`.
-    TildeMarker,
-    /// Goal matching produced by `~`: match `inner` first, then `goal`,
-    /// but preserve capture order as written (`goal` before `inner`).
-    GoalMatch {
-        goal: RegexPattern,
-        inner: RegexPattern,
-        goal_text: String,
-    },
-}
-
-#[derive(Clone)]
-enum RegexQuant {
-    One,
-    ZeroOrMore,
-    OneOrMore,
-    ZeroOrOne,
-    /// `** min..max` — repeat exactly min to max times (max=None means unbounded)
-    Repeat(usize, Option<usize>),
-    /// `** {code}` — repeat count determined at runtime by evaluating code block
-    RepeatCode(String),
-}
-
-#[derive(Clone)]
-struct CharClass {
-    negated: bool,
-    items: Vec<ClassItem>,
-}
-
-#[derive(Clone)]
-enum ClassItem {
-    Range(char, char),
-    Char(char),
-    Digit,
-    NegDigit,
-    Word,
-    NegWord,
-    Space,
-    NegSpace,
-    HorizSpace,
-    NegHorizSpace,
-    VertSpace,
-    NegVertSpace,
-    NotNewline,
-    NamedBuiltin(String),
-    UnicodePropItem { name: String, negated: bool },
 }
 
 /// Entry in the callframe stack, tracking state for each call frame.
